Adding "Hybrid Cryptography" to the article
Hybrid Cryptography is a hybrid based encryption with both Asymmetric and Symmetric Key Encryptions. This is achieved by using public key encryption to protect the message by adding an unsystematic layer over the symmetic key encryption. Next, the public key encryption technique is used to decode the symmetic key. After the symmetic key is uncovered, the message is ready to be decrypted to plaintext. This method is shown to be quite useful in providing extra security in cryptographic messages. 

Link: (https://www.techopedia.com/definition/1779/hybrid-encryption)